+++ src/Cache.php
@@ -8,7 +8,7 @@ class Cache {
 	
 	/**
 	 * __construct
-	 * @version 1.0 (2023-05-03)
+	 * @version 1.0.1 (2024-07-29)
 	 */
 	public function __construct(){
 		$this->cacheDir =
@@ -16,47 +16,48 @@ class Cache {
 			dirname(
 				__DIR__,
 				3
-			) .
-			'/cache/ddRunSnippets'
+			)
+			. '/cache/ddRunSnippets'
 		;
 		
 		if (!is_dir($this->cacheDir)){
 			\DDTools\FilesTools::createDir([
-				'path' => $this->cacheDir
+				'path' => $this->cacheDir,
 			]);
 		}
 	}
 	
 	/**
 	 * getCache
-	 * @version 1.0 (2023-05-03)
+	 * @version 2.1 (2024-07-29)
 	 * 
-	 * @param $params {stdClass|arrayAssociative} — Parameters, the pass-by-name style is used. @required
-	 * @param $params->docId {integer} — Document ID related to cache. @required
-	 * @param $params->name {string} — Unique cache name for the document. @required
-	 * @param $params->data {string|array|stdClass} — Data to save. @required
+	 * @param $params {stdClass|arrayAssociative} — Parameters, the pass-by-name style is used.
+	 * @param $params->resourceId {integer} — Document ID related to cache.
+	 * @param $params->name {string} — Unique cache name for the document.
+	 * @param $params->data {string|array|stdClass} — Data to save.
+	 * @param [$params->prefix='doc'] {string} — Cache prefix.
 	 * 
 	 * @return {void}
 	 */
-	public function createCache($params) :void {
+	public function createCache($params): void {
 		$params = (object) $params;
 		
 		//str|obj|arr
 		$dataType =
-			is_object($params->data) ?
-			'obj' :
-			(
-				is_array($params->data) ?
-				'arr' :
+			is_object($params->data)
+			? 'obj'
+			: (
+				is_array($params->data)
+				? 'arr'
 				//All other types are considered as string (because of this we don't use the gettype function)
-				'str'
+				: 'str'
 			)
 		;
 		
 		if ($dataType != 'str'){
 			$params->data = \DDTools\ObjectTools::convertType([
 				'object' => $params->data,
-				'type' => 'stringJsonAuto'
+				'type' => 'stringJsonAuto',
 			]);
 		}
 		
@@ -66,20 +67,21 @@ class Cache {
 			$this->buildCacheFilePath($params),
 			//Cache content
 			(
-				$this->contentPrefix .
-				$dataType .
-				$params->data
+				$this->contentPrefix
+				. $dataType
+				. $params->data
 			)
 		);
 	}
 	
 	/**
 	 * getCache
-	 * @version 1.0 (2023-05-03)
+	 * @version 2.1 (2024-07-29)
 	 * 
-	 * @param $params {stdClass|arrayAssociative} — Parameters, the pass-by-name style is used. @required
-	 * @param $params->docId {integer} — Document ID related to cache. @required
-	 * @param $params->name {string} — Unique cache name for the document. @required
+	 * @param $params {stdClass|arrayAssociative} — Parameters, the pass-by-name style is used.
+	 * @param $params->resourceId {integer} — Document ID related to cache.
+	 * @param $params->name {string} — Unique cache name for the document.
+	 * @param [$params->prefix='doc'] {string} — Cache prefix.
 	 * 
 	 * @return {null|string|array|stdClass} — `null` means cache is not exist.
 	 */
@@ -112,9 +114,10 @@ class Cache {
 				$result = \DDTools\ObjectTools::convertType([
 					'object' => $result,
 					'type' =>
-						$dataType == 'obj' ?
-						'objectStdClass' :
-						'objectArray'
+						$dataType == 'obj'
+						? 'objectStdClass'
+						: 'objectArray'
+					,
 				]);
 			}
 		}
@@ -124,38 +127,40 @@ class Cache {
 	
 	/**
 	 * clearCache
-	 * @version 1.0 (2023-05-03)
+	 * @version 2.1 (2024-07-29)
 	 * 
 	 * @param Clear cache files for specified document or every documents.
 	 * 
-	 * @param $params {stdClass|arrayAssociative} — Parameters, the pass-by-name style is used. Default: —.
-	 * @param $params->docId {integer|null} — Document ID related to cache. Default: null (cache of all docs will be cleared).
+	 * @param [$params] {stdClass|arrayAssociative} — Parameters, the pass-by-name style is used.
+	 * @param $params->resourceId {integer|null} — Document ID related to cache. Default: null (cache of all docs will be cleared).
+	 * @param [$params->prefix='doc'] {string} — Cache prefix.
 	 * 
 	 * @return {void}
 	 */
-	public function clearCache($params) :void {
-		\DDTools\ObjectTools::extend([
+	public function clearCache($params = []): void {
+		$params = \DDTools\ObjectTools::extend([
 			'objects' => [
 				(object) [
-					'docId' => null
+					'resourceId' => null,
+					'prefix' => 'doc',
 				],
-				$params
-			]
+				$params,
+			],
 		]);
 		
 		//Clear all cache
-		if (empty($params->docId)){
+		if (empty($params->resourceId)){
 			\DDTools\FilesTools::removeDir($this->cacheDir);
 		//Clear cache for specified documents
 		}else{
 			$files = glob(
-				$this->cacheDir .
-				'/doc' . $params->docId . '-*.php'
+				$this->cacheDir
+				. '/' . $params->prefix . $params->resourceId . '-*.php'
 			);
 			
 			foreach (
-				$files as
-				$filepath
+				$files
+				as $filepath
 			){
 				unlink($filepath);
 			}
@@ -164,20 +169,28 @@ class Cache {
 	
 	/**
 	 * buildCacheFilePath
-	 * @version 1.0 (2023-05-03)
+	 * @version 2.1 (2024-07-29)
 	 * 
-	 * @param $params {stdClass|arrayAssociative} — Parameters, the pass-by-name style is used. @required
-	 * @param $params->docId {integer} — Document ID related to cache. @required
-	 * @param $params->name {string} — Unique cache name for the document. @required
+	 * @param $params {stdClass|arrayAssociative} — Parameters, the pass-by-name style is used.
+	 * @param $params->resourceId {integer} — Document ID related to cache.
+	 * @param $params->name {string} — Unique cache name for the document.
+	 * @param [$params->prefix='doc'] {string} — Cache prefix.
 	 * 
 	 * @return {string}
 	 */
-	private function buildCacheFilePath($params) :string {
-		$params = (object) $params;
+	private function buildCacheFilePath($params): string {
+		$params = \DDTools\ObjectTools::extend([
+			'objects' => [
+				(object) [
+					'prefix' => 'doc',
+				],
+				$params,
+			],
+		]);
 		
 		return
-			$this->cacheDir .
-			'/doc' . $params->docId . '-' . $params->name . '.php'
+			$this->cacheDir
+			. '/' . $params->prefix . $params->resourceId . '-' . $params->name . '.php'
 		;
 	}
 }